feature update for new vexml code and implementation of this code for the navigate tab on the gui. this tab is now on the main veconductor interface. I also removed all the winsock2 headers as that compile problem was caused by a conflict between wx and tao.
my last commit failed in the middle so here is the rest of the code.

+++ VE_TextureBased/cfdTextureManager.h
@@ -85,7 +85,7 @@ public:
    unsigned char* getNextField(/*int plusNeg*/);
    unsigned int getNextFrame();
    //get the number of vector fields
-   int numberOfFields(){return _dataFields.size();}
+   int numberOfFields(){return static_cast< int >( _dataFields.size() );}
 
    //the resolution of the fields
    int* fieldResolution(){return _resolution;}
